Updated ITER to long

+++ src/Base/FractalParam.java
@@ -6,7 +6,7 @@ import java.util.List;
 public enum FractalParam {
     XRES (Integer.class, Arrays.asList("XRES", "x")),
     YRES (Integer.class, Arrays.asList("YRES", "y")),
-    ITER (Integer.class, Arrays.asList("ITER", "i")),
+    ITER (Long.class, Arrays.asList("ITER", "i")),
     VAR0 (Double.class, Arrays.asList("VAR0", "v0", "w0", "linear"), new Color(1.0,0.0,0.0)),
     VAR1 (Double.class, Arrays.asList("VAR1", "v1", "w1", "sinusodial"), new Color(0.0,1.0,0.0)),
     VAR2 (Double.class, Arrays.asList("VAR2", "v1", "w2", "spherical"), new Color(0.0,0.0,1.0));
